part3/requirements.txt
flask
pandas
numpy
python-dotenv
gunicorn
resend